Switch from an interval between refreshes to specific times
Previously, refreshes occurred after a fixed interval (say 30
minutes). This was fine but caused the times displayed on screen
to be odd numbers of minutes and drift from time to time.

Now we have an array of "minutes past the hour" at which the
display should update. This results in nice looking times and
something that's more predictable for the human looking at it.